refactor out proven data
