+++ src/main/scala/edg_ide/swing/EdgirLibraryTreeTableModel.scala
@@ -34,14 +34,6 @@ trait EdgirLibraryNodeBase {  // abstract base class for tree node model
 }
 
 
-// Data holder for a Block's proven status
-sealed trait ProvenNodeBase
-
-object EmptyProven extends ProvenNodeBase {
-  override def toString = ""
-}
-
-
 class EdgirLibraryNode(project: Project, library: edg.wir.Library) extends EdgirLibraryNodeBase {
   // This (root) node properties
   //
@@ -65,28 +57,6 @@ class EdgirLibraryNode(project: Project, library: edg.wir.Library) extends Edgir
     override lazy val children: Seq[EdgirLibraryNodeBase] = {
       paths.map { childPath => new BlockNode(childPath, library.allBlocks(childPath), root) }
           .sortBy(_.toString)
-    }
-  }
-
-  class BlockProven(path: ref.LibraryPath,
-                    val records: BlockProvenRecords) extends ProvenNodeBase {
-    override lazy val toString: String = if (records.isEmpty) {
-      ""
-    } else {
-      records.getDataOfStatus(records.getLatestStatus).flatMap { case (design, records) => records }.size.toString
-    }
-
-    lazy val htmlDescription = {
-      val dataFormatted = records.data.flatMap { case ((file, version), records) =>
-        records.groupBy(_._1.status).map { case (status, records) =>
-          val statusColor = SwingHtmlUtil.colorToHtml(ProvenStatus.colorOf(status))
-          s"""<font style="color:$statusColor;"><b>$status</b>: ${file.getName}@${version.substring(0, 7)} (${records.size})</font>"""
-        }
-      }
-
-      f"""<b>${path.toSimpleString}</b><hr>
-          ${dataFormatted.mkString("<br/>")}
-          """
     }
   }
